Completely change over to self-hosted charts
