+++ app/views/beers/index/_beer.html.erb
@@ -1,7 +1,7 @@
 <div class="flex flex-col rounded-lg shadow-lg overflow-hidden">
   <div class="flex-shrink-0">
     <div class="relative h-48 w-full flex justify-center items-center">
-      <%= render partial: 'beers/index/beer_chart', locals: { beer: beer } %>
+      <%= render partial: 'beers/shared/beer_chart', locals: { beer: beer } %>
     </div>
   </div>
   <div class="flex-1 bg-white p-6 flex flex-col justify-between z-10">
