--- app/views/beers/show/_beer_chart.html.erb
@@ -1,40 +1,0 @@
-<canvas id="beer-show-<%= beer.id %>" class="my-auto"></canvas>
-<script type="text/javascript">
-  var ctx = document.getElementById('beer-show-<%= beer.id %>').getContext('2d');
-  var data = {
-    datasets: [{
-      data: [
-        <%= beer.time_after_start_before_bottle.to_i %>,
-        <%= beer.time_until_bottle.to_i %>,
-        <%= beer.time_after_bottle_before_ready.to_i %>,
-        <%= beer.time_until_ready.to_i %>,
-      ],
-      backgroundColor: [
-        'forestgreen',
-        'gainsboro',
-        'forestgreen',
-        'gainsboro'
-      ]
-    }],
-    // These labels appear in the legend and in the tooltips when hovering different arcs
-    labels: [
-      '<%= beer.bottled? ? "Beer is bottled" : "Days elapsed since start" %>',
-      'Days until bottled',
-      '<%= beer.brewed? ? "Beer is brewed" : "Days elapsed since bottled" %>',
-      'Days until brewed',
-    ]
-  };
-  var options = {
-    rotation: -Math.PI,
-    circumference: Math.PI,
-    legend: {
-      display: false,
-    },
-    aspectRatio: 2.30,
-  };
-  new Chart(ctx, {
-    type: 'doughnut',
-    data: data,
-    options: options
-  });
-</script>
